Displaying basket items on the basket page using new controller and view
